// packages/@ros-cdk/services/rds/lib/rds.generated.ts
export interface RosAccountPrivilegeProps {

    /**
     * @Property accountName: RDS account name.
     */
    readonly accountName: string;

    /**
     * @Property accountPrivilege: RDS account privilege
     */
    readonly accountPrivilege: string;

    /**
     * @Property dbInstanceId: RDS instance ID.
     */
    readonly dbInstanceId: string;

    /**
     * @Property dbName: RDS database name
     */
    readonly dbName: string;
}
// ...
export class RosAccountPrivilege extends ros.RosResource {
    /**
     * The resource type name for this resource class.
     */
    public static readonly ROS_RESOURCE_TYPE_NAME = "ALIYUN::RDS::AccountPrivilege";

    /**
     * A factory method that creates a new instance of this class from an object
     * containing the properties of this ROS resource.
     */

    public enableResourcePropertyConstraint: boolean;


    /**
     * @Property accountName: RDS account name.
     */
    public accountName: string;

    /**
     * @Property accountPrivilege: RDS account privilege
     */
    public accountPrivilege: string;

    /**
     * @Property dbInstanceId: RDS instance ID.
     */
    public dbInstanceId: string;

    /**
     * @Property dbName: RDS database name
     */
    public dbName: string;

    /**
     * Create a new `ALIYUN::RDS::AccountPrivilege`.
     *
     * @param scope - scope in which this resource is defined
     * @param id    - scoped id of the resource
     * @param props - resource properties
     */
    constructor(scope: ros.Construct, id: string, props: RosAccountPrivilegeProps, enableResourcePropertyConstraint: boolean) {
        super(scope, id, { type: RosAccountPrivilege.ROS_RESOURCE_TYPE_NAME, properties: props });

        this.enableResourcePropertyConstraint = enableResourcePropertyConstraint;
        this.accountName = props.accountName;
        this.accountPrivilege = props.accountPrivilege;
        this.dbInstanceId = props.dbInstanceId;
        this.dbName = props.dbName;
    }


    protected get rosProperties(): { [key: string]: any }  {
        return {
            accountName: this.accountName,
            accountPrivilege: this.accountPrivilege,
            dbInstanceId: this.dbInstanceId,
            dbName: this.dbName,
        };
    }
    protected renderProperties(props: {[key: string]: any}): { [key: string]: any }  {
        return rosAccountPrivilegePropsToRosTemplate(props, this.enableResourcePropertyConstraint);
    }
}